refactor: OptionLoginAsyncExtension
